README.md
This repository contains the software components for an autonomous robot navigation and control system, designed for an ESP32 microcontroller and a simulated environment (Webots), complemented by a real-time visualization dashboard. The system integrates modules for network connectivity,  sensor processing and motor control, advanced pathfinding, and a visualization interface for monitoring robot operations.